refactor: cleanup in context menu, tooltip, toast
