+++ src/lib/context-menu/ContextSubMenuContent.svelte
@@ -1,24 +1,13 @@
 <script lang="ts">
-  import {
-    highlightMenuItem,
-    isMenuActionActivationKey,
-    shouldCloseMenuOnAction,
-  } from './behavior'
-  import { fade, scale } from 'svelte/transition'
+  import { scale } from 'svelte/transition'
   import { cubicOut } from 'svelte/easing'
   import { MENU_MAX_HEIGHT, MENU_WIDTH } from './const'
-  import {
-    type MenuFlyoutItem,
-    type MenuInteractiveItem,
-    isMenuComponentItem,
-    isMenuDivider,
-    isMenuFlyoutItem,
-  } from './types'
+  import { type MenuFlyoutItem, isMenuComponentItem } from './types'
   import { useContextMenu } from './contextMenuState.svelte'
+  import { portal } from '$lib/shared/placement'
+  import MenuList from './MenuList.svelte'
 
   const contextMenuState = useContextMenu()
-  import { portal } from '$lib/shared/placement'
-  import ContextSubMenu from './ContextSubMenu.svelte'
 
   interface PositionAction {
     update?: () => void
@@ -40,32 +29,6 @@
       ? item.componentWidth
       : MENU_WIDTH
   )
-
-  let activeChildLabel = $state<string | null>(null)
-
-  const handleChildAction = (child: MenuInteractiveItem) => {
-    if (child.disabled) return
-
-    if (child.value !== undefined) {
-      child.onAction?.(child.value)
-    } else {
-      child.onAction?.()
-    }
-
-    highlightMenuItem(item.children, child.label)
-
-    if (shouldCloseMenuOnAction(child)) {
-      contextMenuState.reset()
-    }
-  }
-
-  const handleKeydown = (e: KeyboardEvent, child: MenuInteractiveItem) => {
-    if (isMenuActionActivationKey(e.key)) {
-      e.preventDefault()
-      e.stopPropagation()
-      handleChildAction(child)
-    }
-  }
 </script>
 
 <div
@@ -102,51 +65,7 @@
         />
       </div>
     {:else if item.children}
-      <ul>
-        {#each item.children as child}
-          {#if isMenuDivider(child)}
-            <li class="divider" role="presentation"></li>
-          {:else if isMenuFlyoutItem(child)}
-            <ContextSubMenu
-              item={child}
-              siblings={item.children}
-              parentZIndex={parentZIndex + 1}
-              isOpen={activeChildLabel === child.label}
-              onToggle={() =>
-                (activeChildLabel =
-                  activeChildLabel === child.label
-                    ? null
-                    : (child.label ?? null))}
-            />
-          {:else}
-            <li>
-              <button
-                role="menuitem"
-                class:selected={child.isHighlighted}
-                disabled={child.disabled}
-                onclick={e => {
-                  e.stopPropagation()
-                  handleChildAction(child)
-                }}
-                onkeydown={e => handleKeydown(e, child)}
-              >
-                {#if child.icon}
-                  {@const ChildIcon = child.icon}
-                  <ChildIcon size={'1em'} strokeWidth={1} />
-                {/if}
-                {#if child.detail}
-                  <span class="item-body">
-                    <span class="item-label">{child.label}</span>
-                    <span class="item-detail">{child.detail}</span>
-                  </span>
-                {:else}
-                  {child.label}
-                {/if}
-              </button>
-            </li>
-          {/if}
-        {/each}
-      </ul>
+      <MenuList items={item.children} parentZIndex={parentZIndex + 1} />
     {/if}
   </div>
 </div>
@@ -168,83 +87,6 @@
     overflow-y: auto;
     overflow-x: hidden;
     padding: 4px 0;
-  }
-
-  li.divider {
-    height: 1px;
-    background: var(--c-grey);
-    margin: 4px 0;
-  }
-
-  ul {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-  }
-
-  button {
-    background: none;
-    border: none;
-    padding: 6px 12px;
-    font-size: 13px;
-    color: var(--c-text);
-    cursor: pointer;
-    text-align: left;
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    transition: background var(--transition-fast) ease;
-    border-radius: var(--rounded-md);
-    margin: 0 4px;
-    width: calc(100% - 8px);
-  }
-
-  button:focus-visible {
-    outline: 2px solid var(--c-brand);
-    outline-offset: -2px;
-  }
-
-  button:hover {
-    background: var(--c-lightgrey);
-    color: var(--c-black);
-  }
-
-  button.selected {
-    background: color-mix(in srgb, var(--c-brand) 6%, var(--c-white));
-    color: var(--c-brand);
-    font-weight: 500;
-  }
-
-  button.selected:hover {
-    background: color-mix(in srgb, var(--c-brand) 10%, var(--c-white));
-  }
-
-  .item-body {
-    display: flex;
-    flex-direction: column;
-    gap: 2px;
-    min-width: 0;
-    flex: 1;
-  }
-
-  .item-label {
-    line-height: 1.3;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-
-  .item-detail {
-    font-size: 10px;
-    color: var(--c-darkgrey);
-    line-height: 1.2;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-
-  button.selected .item-detail {
-    color: color-mix(in srgb, var(--c-brand) 70%, var(--c-darkgrey));
   }
 
   .custom-component-wrap {
