+++ src/lib/plots/aoi-stream/components/AoiStreamPlotFigure.svelte
@@ -623,7 +623,6 @@
     }
 
     return {
-      tooltipId: 'aoi-stream-bin-tooltip',
       content,
       anchorX: mx,
       anchorY: my,
