+++ src/lib/plots/scanpath-similarity/components/ScangraphFigure.svelte
@@ -554,7 +554,6 @@
    * schedules no repaint.
    */
   const EMPTY_HIT: FrameHit<ScangraphHover> = {
-    tooltipId: 'scangraph-tooltip',
     content: [],
     anchorX: 0,
     anchorY: 0,
@@ -572,7 +571,6 @@
         for (let i = 1; i < connectionItems.length; i++) content.push(connectionItems[i])
       }
       return {
-        tooltipId: 'scangraph-tooltip',
         content,
         anchorX: node.x + 10,
         anchorY: node.y,
@@ -591,7 +589,6 @@
     const s = layoutResult.nodes[link.source]
     const t = layoutResult.nodes[link.target]
     return {
-      tooltipId: 'scangraph-tooltip',
       content: [
         { key: 'Participants', value: `${clipLabel(s.label)} ↔ ${clipLabel(t.label)}` },
         { key: 'Similarity', value: link.value.toFixed(3) },
