+++ tests/selectContextMenuPrimitives.test.ts
@@ -2,7 +2,6 @@ import { beforeEach, describe, expect, it } from 'vitest'
 import type { Component } from 'svelte'
 import {
   highlightMenuItem,
-  isMenuActionActivationKey,
   isOwnedContextMenuState,
   shouldCloseMenuOnAction,
 } from '../src/lib/context-menu/behavior'
@@ -61,7 +60,7 @@ describe('select and context-menu primitives', () => {
     expect(getSelectLabel(undefined, [])).toBe('')
   })
 
-  it('keeps menu component items on the flyout path and exposes activation helpers', () => {
+  it('keeps menu component items on the flyout path', () => {
     const componentItem = createMenuComponentItem<Record<string, unknown>>({
       label: 'Custom',
       value: 'custom',
@@ -76,9 +75,6 @@ describe('select and context-menu primitives', () => {
     expect(isMenuComponentItem(componentItem)).toBe(true)
     expect(isMenuFlyoutItem(componentItem)).toBe(true)
     expect(isMenuFlyoutItem(flyoutItem)).toBe(true)
-    expect(isMenuActionActivationKey('Enter')).toBe(true)
-    expect(isMenuActionActivationKey(' ')).toBe(true)
-    expect(isMenuActionActivationKey('Escape')).toBe(false)
   })
 
   it('tracks menu ownership, highlighting, and close-on-action semantics', () => {
@@ -97,7 +93,6 @@ describe('select and context-menu primitives', () => {
     }
 
     contextMenuState.current = {
-      visible: true,
       items,
       x: 10,
       y: 20,
